adds impersonate function

+++ contracts/interfaces/IExafin.sol
@@ -7,11 +7,41 @@ struct Pool {
 }
 
 interface IExafin {
-    function rateForSupply(uint256 amount, uint256 maturityDate) external view returns (uint256, Pool memory);
-    function rateToBorrow(uint256 amount, uint256 maturityDate) external view returns (uint256, Pool memory);
-    function borrow(address to, uint256 amount, uint256 maturityDate) external;
-    function supply(address from, uint256 amount, uint256 maturityDate) external;
+    function rateForSupply(uint256 amount, uint256 maturityDate)
+        external
+        view
+        returns (uint256, Pool memory);
+
+    function rateToBorrow(uint256 amount, uint256 maturityDate)
+        external
+        view
+        returns (uint256, Pool memory);
+
+    function borrow(
+        address to,
+        uint256 amount,
+        uint256 maturityDate
+    ) external;
+
+    function supply(
+        address from,
+        uint256 amount,
+        uint256 maturityDate
+    ) external;
+
     function tokenName() external view returns (string calldata);
-    function getAccountSnapshot(address who, uint timestamp) external view returns (uint, uint, uint);
-    function getTotalBorrows(uint256 maturityDate) external view returns (uint);
+
+    function getAccountSnapshot(address who, uint256 timestamp)
+        external
+        view
+        returns (
+            uint256,
+            uint256,
+            uint256
+        );
+
+    function getTotalBorrows(uint256 maturityDate)
+        external
+        view
+        returns (uint256);
 }
